# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth": {
            "post": {
                "description": "Аутентификация пользователя",
                "produces": [
                    "application/json"
                ],
                "summary": "Аутентификация",
                "parameters": [
                    {
                        "description": "Message content",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.requestAuth"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.responseTokens"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.responseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.responseError"
                        }
                    }
                }
            }
        },
        "/api/refresh": {
            "get": {
                "description": "Обновление токена доступа пользователя",
                "produces": [
                    "application/json"
                ],
                "summary": "Обновление токена",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Authorization-Access",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Refresh token",
                        "name": "Authorization-Refresh",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/server.responseTokens"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.responseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.responseError"
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "description": "Зарегистрировать нового пользователя",
                "produces": [
                    "application/json"
                ],
                "summary": "Регистрация",
                "parameters": [
                    {
                        "description": "Message content",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.requestNewUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.responseNewUser"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.responseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.responseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.requestAuth": {
            "type": "object",
            "required": [
                "guid",
                "password"
            ],
            "properties": {
                "guid": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "server.requestNewUser": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "server.responseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "server.responseNewUser": {
            "type": "object",
            "properties": {
                "guid": {
                    "type": "string"
                }
            }
        },
        "server.responseTokens": {
            "type": "object",
            "properties": {
                "access": {
                    "type": "string"
                },
                "refresh": {
                    "type": "string"
                }
            }
        }
    }
}